Small changes
Fixed test, and database

+++ preprocessor/src/main.rs
@@ -249,8 +249,8 @@ fn feature_extraction_to_database(
         .to_db_type(image_id)
         .into_iter()
         .map(|keypoint| DbKeypoints {
-            x_coord: keypoint.x_coord + (column * tile_size * 2_u64.pow(lod as u32)) as f64,
-            y_coord: keypoint.y_coord + (row * tile_size * 2_u64.pow(lod as u32)) as f64,
+            x_coord: keypoint.x_coord + (column * tile_size * 2_u64.pow(lod as u32)) as f32,
+            y_coord: keypoint.y_coord + (row * tile_size * 2_u64.pow(lod as u32)) as f32,
             ..keypoint
         })
         .collect();
